De-paginate when checking for existing asset
The overwrite logic for GitHub Releases first checks to see if an asset
with the given name already exists, and if so deletes it to make way for
the new upload.  If, for some reason, a match is overlooked, the asset
is not deleted, and the upload errors with 422 Validation Failed, code
"already_exists" and field "name".

The "some reason" in this case was that only the first page of assets
are returned in-line with the releases() call.  Using release_assets()
directly will trigger the paginate logic in Octokit and fix the problem.